+++ strato/indexer/slipstream/src/Slipstream/Processor.hs
@@ -128,6 +128,7 @@ processedContract ABIID {..} state AggregateAction {..} =
     { address = actionAccount ^. accountAddress,
       codehash = actionCodeHash,
       creator = actionCreator,
+      cc_creator = actionCCCreator,
       root = actionRoot,
       application = actionApplication,
       contractName = aiName,
@@ -169,8 +170,8 @@ rowToCollections row = do
         _ -> []
   return $ (Map.fromList $ newState)
 
-processedContractToProcessedCollectionRows :: MonadIO m => Map.Map Text Value -> [Text] -> AggregateAction -> ABIID -> m [ProcessedCollectionRow]
-processedContractToProcessedCollectionRows state mapAndArrayNames row abiid = do
+processedContractToProcessedCollectionRows :: MonadIO m => Map.Map Text Value -> [Text] -> AggregateAction -> ABIID -> Maybe Text ->m [ProcessedCollectionRow]
+processedContractToProcessedCollectionRows state mapAndArrayNames row abiid cregator = do
   let valueCollectionsMap = Map.filter (\value -> case value of 
                                                       ValueMapping _ -> True 
                                                       ValueArrayFixed _ _ -> True 
@@ -190,11 +191,11 @@ processedContractToProcessedCollectionRows state mapAndArrayNames row abiid = do
           processRecord (mName, value) = 
             case value of
               Left theMap -> 
-                map (\(k, v) -> processedCollectionRow mName (T.pack "Mapping") row abiid (SimpleValue k) v) (Map.toList theMap)
+                map (\(k, v) -> processedCollectionRow mName (T.pack "Mapping") row abiid cregator (SimpleValue k) v) (Map.toList theMap)
               Right (Left arrayValues) -> 
-                map (processArrayFixed mName row abiid) (zip [0..] arrayValues)
+                map (processArrayFixed mName row abiid cregator) (zip [0..] arrayValues)
               Right (Right intMapValues) -> 
-                map (processArrayDynamic mName row abiid) (I.toList intMapValues)
+                map (processArrayDynamic mName row abiid cregator) (I.toList intMapValues)
       return result
 
 rowToHistories ::
@@ -213,12 +214,13 @@ rowToHistories _ abiId actions cont oldState = do
     newMap <- gets Map.fromList
     return $ processedContract abiId newMap hRow
 
-processedCollectionRow :: Text -> Text -> AggregateAction -> ABIID -> Value -> Value -> ProcessedCollectionRow
-processedCollectionRow collection ttype AggregateAction {..} ABIID {..} k v =
+processedCollectionRow :: Text -> Text -> AggregateAction -> ABIID -> Maybe Text -> Value -> Value -> ProcessedCollectionRow
+processedCollectionRow collection ttype AggregateAction {..} ABIID {..} cregator k v =
   ProcessedCollectionRow
     { address = actionAccount ^. accountAddress,
       codehash = actionCodeHash,
       creator = actionCreator,
+      cc_creator = cregator,
       root = actionRoot,
       application = actionApplication,
       contractname = aiName,
@@ -233,13 +235,13 @@ processedCollectionRow collection ttype AggregateAction {..} ABIID {..} k v =
       collectionDataValue = v
     }
 
-processArrayFixed :: Text -> AggregateAction -> ABIID -> (Int, Value) -> ProcessedCollectionRow
-processArrayFixed mName row abiid (index, value) =
-  processedCollectionRow mName (T.pack "Array") row abiid (SimpleValue (ValueInt False Nothing (fromIntegral index))) value
+processArrayFixed :: Text -> AggregateAction -> ABIID -> Maybe Text -> (Int, Value) -> ProcessedCollectionRow
+processArrayFixed mName row abiid cregator (index, value) =
+  processedCollectionRow mName (T.pack "Array") row abiid cregator (SimpleValue (ValueInt False Nothing (fromIntegral index))) value
 
-processArrayDynamic :: Text -> AggregateAction -> ABIID -> (Int, Value) -> ProcessedCollectionRow
-processArrayDynamic mName row abiid (index, value) =
-  processedCollectionRow mName (T.pack "Array") row abiid (SimpleValue (ValueInt False Nothing (fromIntegral index))) value
+processArrayDynamic :: Text -> AggregateAction -> ABIID -> Maybe Text -> (Int, Value) -> ProcessedCollectionRow
+processArrayDynamic mName row abiid cregator (index, value) =
+  processedCollectionRow mName (T.pack "Array") row abiid cregator (SimpleValue (ValueInt False Nothing (fromIntegral index))) value
 
 -- Prioritizing with-source actions prevents the issue where updates to contracts
 -- at different addresses are lost because the schema has not been seen yet.
@@ -436,7 +438,7 @@ processTheMessages env conn messages = do
                     aiChain = cid
                   }
               cont = error "internal error: contract should be unused for SolidVM"
-          $logDebugLS "Contract name is: " $ T.pack $ show name
+          $logInfoLS "Contract name is: " $ T.pack $ show name
           oldState <- readPreviousSolidVMState g acct
           indexContract <- rowToInsert g abiid row cont oldState
           let fkeysForThisContract = getContractsFromPC indexContract
@@ -446,10 +448,13 @@ processTheMessages env conn messages = do
               collectionNames = mapNames ++ arrNames
               abstracts = actionAbstracts row
           --get columns for abstract table
-          $logDebugLS "abstractColumns" $ T.pack $ "Getting abstract columns from " ++ (show abstracts)
+          $logInfoLS "abstractColumns" $ T.pack $ "Getting abstract columns from " ++ (show abstracts)
           abstractColumns' <- fmap catMaybes . for (Map.toList abstracts) $ \((_, n'), (cr', ap')) -> do
-            let tableName = AbstractTableName cr' ap' n'
+            let cregator = fromMaybe cr' (actionCCCreator row)
+                tableName = AbstractTableName cregator ap' n'
                 tableNameText = tableNameToDoubleQuoteText tableName
+            $logDebugLS "actionCCCreator" $ T.pack (show (actionCCCreator row))
+            $logDebugLS "cregator" $ T.pack (show cregator)
             $logInfoS "Row will be inserted into abstract table: " tableNameText
             mCols <- getTableColumns g tableName
             pure $ (indexContract, fkeysForThisContract, tableNameText, (cr',ap',n'),) . map extractTextInsideQuotes <$> mCols
@@ -459,7 +464,7 @@ processTheMessages env conn messages = do
           stateDiff <- rowToCollections row
           parents' <- pure $ map (\(_,_,_,p ,_)-> p) abstractColumns'
           abstractColumns <- pure $ map (\(a,b,c,_,e) -> (a,b,c,e)) abstractColumns'
-          pCollections <- processedContractToProcessedCollectionRows stateDiff (collectionNames) row abiid --get all collection rows to insert
+          pCollections <- processedContractToProcessedCollectionRows stateDiff (collectionNames) row abiid (actionCCCreator row) --get all collection rows to insert
           pCollectionsWithAbstracts <- pure $ duplicateForParentsAndIncludeOriginal pCollections parents'
           pure . Right $ BatchedInserts (indexContract, fkeysForThisContract) abstractColumns hs pCollectionsWithAbstracts
 
@@ -491,7 +496,7 @@ processTheMessages env conn messages = do
     notifyPostgREST conn
 
   when (length events' > 0) $
-    outputData conn $ insertEventTables g events'
+    outputData conn $ insertEventTables events'
 
   $logInfoS "processTheMessages" . T.pack $ "Inserting " ++ show (length transactionResults) ++ " transaction results"